backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.3
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.4
kafka-python==2.0.2
PyJWT==2.8.0
python-decouple==3.8
Pillow==10.1.0
django-extensions==3.2.3
drf-spectacular==0.26.5
requests==2.31.0
python-dotenv==1.0.0
django-tenant-schemas==1.10.0
django-storages==1.14.2
gunicorn==21.2.0
